+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - Add text and raised button
+
+* Add text and raised button
+
+
 ## [0.0.1] - Initial release
 
 * This is initial release.